backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
aiosqlite==0.20.0
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
python-multipart==0.0.12
websockets==13.1
python-dotenv==1.0.1
